feat: alterar tipo de moment_id para BIGINT na migração de moment_metadatas

+++ src/database/migrations/20240205180203-create-moment_metadatas.js
@@ -11,7 +11,7 @@ module.exports = {
         allowNull: false,
       },
       moment_id: {
-        type: Sequelize.INTEGER(),
+        type: Sequelize.BIGINT(),
         allowNull: false,
         references: {
           model: 'moments',
